+++ .local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/lib/Manager.js
@@ -245,6 +245,38 @@ var Manager = class
         this._settings.connect('changed::panel-icon-size', () => {
             this._applyPanelIconSize(false);
         });
+
+        this._settings.connect('changed::dash-separator', () => {
+            this._applyDashSeparator(false);
+        });
+
+        this._settings.connect('changed::looking-glass-width', () => {
+            this._applyLookingGlassSize(false);
+        });
+
+        this._settings.connect('changed::looking-glass-height', () => {
+            this._applyLookingGlassSize(false);
+        });
+
+        this._settings.connect('changed::osd-position', () => {
+            this._applyOSDPosition(false);
+        });
+
+        this._settings.connect('changed::window-menu-take-screenshot-button', () => {
+            this._applyWindowMenuTakeScreenshotButton(false);
+        });
+
+        this._settings.connect('changed::alt-tab-window-preview-size', () => {
+            this._applyAltTabWindowPreviewSize(false);
+        });
+
+        this._settings.connect('changed::alt-tab-small-icon-size', () => {
+            this._applyAltTabSmallIconSize(false);
+        });
+
+        this._settings.connect('changed::alt-tab-icon-size', () => {
+            this._applyAltTabIconSize(false);
+        });
     }
 
     /**
@@ -303,6 +335,13 @@ var Manager = class
         this._applyPanelIconSize(false);
         this._applyEventsButton(false);
         this._applyCalendar(false);
+        this._applyDashSeparator(false);
+        this._applyLookingGlassSize(false);
+        this._applyOSDPosition(false);
+        this._applyWindowMenuTakeScreenshotButton(false);
+        this._applyAltTabWindowPreviewSize(false);
+        this._applyAltTabSmallIconSize(false);
+        this._applyAltTabIconSize(false);
     }
 
     /**
@@ -361,6 +400,13 @@ var Manager = class
         this._applyPanelIconSize(true);
         this._applyEventsButton(true);
         this._applyCalendar(true);
+        this._applyDashSeparator(true);
+        this._applyLookingGlassSize(true);
+        this._applyOSDPosition(true);
+        this._applyWindowMenuTakeScreenshotButton(true);
+        this._applyAltTabWindowPreviewSize(true);
+        this._applyAltTabSmallIconSize(true);
+        this._applyAltTabIconSize(true);
     }
 
     /**
@@ -1238,6 +1284,135 @@ var Manager = class
             this._api.panelIconSetDefaultSize();
         } else {
             this._api.panelIconSetSize(size);
+        }
+    }
+
+    /**
+     * apply dash separator settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyDashSeparator(forceOriginal)
+    {
+        let status = this._settings.get_boolean('dash-separator');
+
+        if (forceOriginal || status) {
+            this._api.dashSeparatorShow();
+        } else {
+            this._api.dashSeparatorHide();
+        }
+    }
+
+    /**
+     * apply looking glass size settings
+     * 
+     * @param {boolean} forceOriginal force original shell setting
+     * 
+     * @returns {void}
+     */
+    _applyLookingGlassSize(forceOriginal)
+    {
+        let widthSize = this._settings.get_int('looking-glass-width');
+        let heightSize = this._settings.get_int('looking-glass-height');
+
+        if (forceOriginal) {
+            this._api.lookingGlassSetDefaultSize();
+        } else {
+            let width = (widthSize !== 0) ? widthSize / 10 : null;
+            let height = (heightSize !== 0) ? heightSize / 10 : null;
+            this._api.lookingGlassSetSize(width, height);
+        }
+    }
+
+    /**
+     * apply osd position settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyOSDPosition(forceOriginal)
+    {
+        let pos = this._settings.get_int('osd-position');
+
+        if (forceOriginal || pos === 0) {
+            this._api.osdPositionSetDefault();
+        } else {
+            this._api.osdPositionSet(pos - 1);
+        }
+    }
+    
+    /**
+     * apply window menu take screenshot button settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyWindowMenuTakeScreenshotButton(forceOriginal)
+    {
+        let status = this._settings.get_boolean('window-menu-take-screenshot-button');
+
+        if (forceOriginal || status) {
+            this._api.screenshotInWindowMenuShow();
+        } else {
+            this._api.screenshotInWindowMenuHide();
+        }
+    }
+
+    /**
+     * apply alt tab window preview size settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyAltTabWindowPreviewSize(forceOriginal)
+    {
+        let size = this._settings.get_int('alt-tab-window-preview-size');
+
+        if (forceOriginal || size === 0) {
+            this._api.altTabWindowPreviewSetDefaultSize();
+        } else {
+            this._api.altTabWindowPreviewSetSize(size);
+        }
+    }
+
+    /**
+     * apply alt tab small icon size settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyAltTabSmallIconSize(forceOriginal)
+    {
+        let size = this._settings.get_int('alt-tab-small-icon-size');
+
+        if (forceOriginal || size === 0) {
+            this._api.altTabSmallIconSetDefaultSize();
+        } else {
+            this._api.altTabSmallIconSetSize(size);
+        }
+    }
+
+    /**
+     * apply alt tab icon size settings
+     *
+     * @param {boolean} forceOriginal force original shell setting
+     *
+     * @returns {void}
+     */
+    _applyAltTabIconSize(forceOriginal)
+    {
+        let size = this._settings.get_int('alt-tab-icon-size');
+
+        if (forceOriginal || size === 0) {
+            this._api.altTabIconSetDefaultSize();
+        } else {
+            this._api.altTabIconSetSize(size);
         }
     }
 }